scape_data/requirements.txt
selenium>=4.0.0
beautifulsoup4>=4.10.0
webdriver-manager>=3.8.0
lxml
python-dateutil